requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
